celery_with_fastapi/pyproject.toml
[project]
name = "celery-with-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ccxt>=4.3.98",
    "celery>=5.4.0",
    "dependency-injector>=4.42.0",
    "fastapi>=0.114.0",
    "flower>=2.0.1",
    "orjson>=3.10.7",
    "pydantic-settings>=2.4.0",
    "pydantic>=2.9.0",
    "pyyaml>=6.0.2",
    "redis>=5.0.8",
    "sqlalchemy>=2.0.34",
    "uvicorn>=0.30.6",
]
